% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra.R
\name{lotka_volterra}
\alias{lotka_volterra}
\title{Lotka-Volterra Predator-Prey Model}
\usage{
lotka_volterra(time, init_n, init_p, r, c, a, m)
}
\arguments{
\item{time}{Vector of time points at which to evaluate system}

\item{init_n}{Initial abundance of prey population}

\item{init_p}{Initial abundance of predator population}

\item{r}{Prey per-capita growth rate}

\item{c}{Predator consumption efficiency (of turning prey into offspring)}

\item{a}{Predator search/attack efficiency}

\item{m}{Predator mortality rate}
}
\value{
Tidy data frame with three columns:
\tabular{lll}{
  \code{Time} \tab \tab Same values and units as argument \code{time}\cr
  \code{Population} \tab \tab One of 'Prey' or 'Predator'\cr
  \code{Abundance} \tab \tab Abundance of given population at given time\cr
}
}
\description{
Lotka-Volterra Predator-Prey Model
}
\examples{
t <- seq(0, 50, 0.1)
lotka_volterra(time = t, init_n = 50, init_p = 30, r = 0.8, c = 0.04, a = 0.2,
               m = 0.3)
}
