% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra_comp.R
\name{lotka_volterra_comp}
\alias{lotka_volterra_comp}
\title{Lotka-Volterra 2-Species Competition Model}
\usage{
lotka_volterra_comp(time, init_n1, init_n2, r1, r2, k1, k2, a1, a2)
}
\arguments{
\item{time}{Vector of time points at which to evaluate system}

\item{init_n1}{Initial abundance of Population 1}

\item{init_n2}{Initial abundance of Population 2}

\item{r1}{Inherent per-capita growth rate of Population 1}

\item{r2}{Inherent per-capita growth rate of Population 2}

\item{k1}{Carrying capacity of Population 1}

\item{k2}{Carrying capacity of Population 2}

\item{a1}{Effect of Population 1 on growth of Population 2}

\item{a2}{Effect of Population 2 on growth of Population 1}
}
\value{
Tidy data frame with three columns:
\tabular{lll}{
  \code{Time} \tab \tab Same values and units as argument \code{time}\cr
  \code{Population} \tab \tab One of 'Pop. 1' or 'Pop. 2'\cr
  \code{Abundance} \tab \tab Abundance of given population at given time\cr
}
}
\description{
Lotka-Volterra 2-Species Competition Model
}
\examples{
t <- seq(0, 50, 0.1)
lotka_volterra_comp(time = t, init_n1 = 10, init_n2 = 10, r1 = 0.5, r2 = 0.1,
                    k1 = 100, k2 = 100, a1 = -0.001, a2 = -0.4)
}
