% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rps.R
\name{rps}
\alias{rps}
\title{Rock-Paper-Scissors Intransitive Competition Model}
\usage{
rps(time, init_r, init_p, init_s, b)
}
\arguments{
\item{time}{Vector of time points at which to evaluate system}

\item{init_r}{Initial proportion playing strategy Rock}

\item{init_p}{Initial proportion playing strategy Paper}

\item{init_s}{Initial proportion playing strategy Scissors}

\item{b}{Benefit to winner of competition}
}
\value{
Tidy data frame with three columns:
\tabular{lll}{
  \code{Time} \tab \tab Same values and units as argument \code{time}\cr
  \code{Strategy} \tab \tab One of 'Rock', 'Paper', or 'Scissors'\cr
  \code{Proportion} \tab \tab Proportion of population playing given strategy
                              at given time\cr
}
}
\description{
Rock-Paper-Scissors Intransitive Competition Model
}
\examples{
t <- seq(0, 50, 0.1)
rps(time = t, init_r = 0.01, init_p = 0.4, init_s = 0.5, b = 0.6)
}
