% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpslk.R
\name{rpslk}
\alias{rpslk}
\title{Rock-Paper-Scissors-Lizard-Spock Intransitive Competition Model}
\usage{
rpslk(time, init_r, init_p, init_s, init_l, init_k, b)
}
\arguments{
\item{time}{Vector of time points at which to evaluate system}

\item{init_r}{Initial proportion playing strategy Rock}

\item{init_p}{Initial proportion playing strategy Paper}

\item{init_s}{Initial proportion playing strategy Scissors}

\item{init_l}{Initial proportion playing strategy Lizard}

\item{init_k}{Initial proportion playing strategy Spock}

\item{b}{Benefit to winner of competition}
}
\value{
Tidy data frame with three columns:
\tabular{lll}{
  \code{Time} \tab \tab Same values and units as argument \code{time}\cr
  \code{Strategy} \tab \tab One of 'Rock', 'Paper', 'Scissors', 'Lizard', or 'Spock'\cr
  \code{Proportion} \tab \tab Proportion of population playing given strategy at given time\cr
}
}
\description{
Rock-Paper-Scissors-Lizard-Spock Intransitive Competition Model
}
\examples{
t <- seq(0, 50, 0.1)
rpslk(time = t, init_r = 0.02, init_p = 0.02, init_s = 0.03, init_l = 0.9,
      init_k = 0.03, b = 0.7)

}
